refactor: kakao script provider
