+++ logics/providers/KakaoScriptProvider.tsx
@@ -1,0 +1,26 @@
+'use client';
+
+import Script from 'next/script';
+
+export default function KakaoScriptProvider({
+  children
+}: React.PropsWithChildren) {
+  const initKakaoScript = () => {
+    if (window.Kakao) {
+      window.Kakao.init(process.env.NEXT_PUBLIC_KAKAO_JS_KEY);
+    }
+  };
+
+  return (
+    <>
+      {children}
+      <Script
+        defer
+        src="https://t1.kakaocdn.net/kakao_js_sdk/2.3.0/kakao.min.js"
+        integrity="sha384-70k0rrouSYPWJt7q9rSTKpiTfX6USlMYjZUtr1Du+9o4cGvhPAWxngdtVZDdErlh"
+        crossOrigin="anonymous"
+        onLoad={initKakaoScript}
+      />
+    </>
+  );
+}
